feat: create html structure to choose the game option
