+++ app.py
@@ -1,4 +1,5 @@
 from flask import Flask, render_template, request, redirect, url_for
+import ai_bot
 
 app = Flask(__name__)
 
@@ -24,49 +25,41 @@ def check_draw():
     return ' ' not in board
 
 
-
 @app.route('/')
+def choice():
+    return render_template('choice.html')
+#
+#
+# @app.route('/two_players')
+# def two_players():
+#     winner = check_winner()
+#     draw = check_draw()
+#     return render_template('index.html',
+#                            board=board,
+#                            current_player=current_player,
+#                            winner=winner,
+#                            draw=draw)
+#
+# @app.route('/ai_play')
+# def ai_play():
+#     ai_bot.ai_move("X", "O")
+#     winner = check_winner()
+#     draw = check_draw()
+#
+
+@app.route('/play')
 def index():
     winner = check_winner()
     draw = check_draw()
+    if current_player == 'O' and len(ai_bot.create_empty_cells_list(board))>0:
+        cell = ai_bot.ai_move(board, "X", "O")
+        play(cell)
     return render_template('index.html',
                            board=board,
                            current_player=current_player,
                            winner=winner,
                             draw=draw)
 
-# LINEAR SEARCH
-# def ai_move(symbol):
-#     win_conditions = [
-#         (0, 1, 2), (3, 4, 5), (6, 7, 8),  # Horizontal
-#         (0, 3, 6), (1, 4, 7), (2, 5, 8),  # Vertical
-#         (0, 4, 8), (2, 4, 6)  # Diagonal
-#     ]
-#     # initialize a position int which will be used to get an index
-#     # on the board in case there is a winning move to be made
-#     position = -1
-#     # go through the winning conditions and compare in pairs,
-#     # if there are 2 symbols in a condition and an empty spot, the empty spot
-#     # will be the position to take for the next move
-#     for condition in win_conditions:
-#         if board[condition[0]] == board[condition[1]] != ' ' and board[condition[2]] == ' ':
-#             position = condition[2]
-#         elif board[condition[0]] == board[condition[2]] != ' ' and board[condition[1]] == ' ':
-#             position = condition[1]
-#         elif board[condition[1]] == board[condition[2]] != ' ' and board[condition[0]] == ' ':
-#             position = condition[0]
-#     if position != -1:
-#         board[position] = symbol
-
-
-# @app.route('/play_ai/<int:cell>')
-# def play_ai(cell):
-#     global current_player
-#     if board[cell] == ' ':
-#         board[cell] = current_player
-#         if not check_winner():
-#             ai_move('O')
-#     return redirect(url_for('index'))
 
 @app.route('/play/<int:cell>')
 def play(cell):
@@ -88,4 +81,4 @@ def reset():
 
 
 if __name__ == '__main__':
-    app.run(debug=True)
+    app.run(debug=True, port="5050")
