+++ templates/index.html
@@ -12,10 +12,6 @@
   <body>
     <div class="container">
       <h1>Tic Tac Toe</h1>
-      <div class="choice">
-           <button id="two_players">2 players</button>
-           <button id="ai_bot">vs AI</button>
-      </div>
       <div class="board">
         {% for index in range(9) %}
         <a href="{{ url_for('play', cell=index) }}" class="cell"
